Add boomerang animator and organize files
Added the sprite class for the boomerang and organized all graphics related scripts into a folder
