+++ Command/MovingDownCommand.cs
@@ -1,5 +1,4 @@
 ﻿using LegendOfZelda.Interfaces;
-using sprint0;
 using System;
 using System.Collections.Generic;
 using System.Linq;
