+++ Graphics/StabLeftSprite.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using sprint0;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,19 +8,19 @@ using System.Threading.Tasks;
 
 namespace LegendOfZelda
 {
-    public class StabUpSprite : AnimatedSprite
+    public class StabLeftSprite : AnimatedSprite
     {
 
-        public StabUpSprite(Texture2D texture, Rectangle[] frames, SpriteEffects effect, int drawFramesPerAnimFrame, int scale) : base(texture, frames, effect, drawFramesPerAnimFrame, scale)
+        public StabLeftSprite(Texture2D texture, Rectangle[] frames, SpriteEffects effect, int drawFramesPerAnimFrame, int scale) : base(texture, frames, effect, drawFramesPerAnimFrame, scale)
         {
 
         }
 
         protected override void DrawSprite()
         {
-            Vector2 adjustedPos = new Vector2(pos.X, pos.Y + (16 - frames[frame].Height) * scale);
+            Vector2 adjustedPos = new Vector2(pos.X + (16 - frames[frame].Width) * scale, pos.Y);
             spriteBatch.Draw(texture, adjustedPos, frames[frame], color, 0, Vector2.Zero, scale, effect, 1);
         }
-            
+
     }
 }
